pyproject.toml
[project]
name = "novel-studio"
version = "0.1.0"
description = "一句话生成完整小说：多层状态机 + Multi-Head Audit + 验收门"
requires-python = ">=3.12"
authors = [{name = "giyojisan", email = "giyojisan@gmail.com"}]

dependencies = [
    "pydantic>=2.9",
    "rich>=13.0",
    "python-slugify>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[project.scripts]
novel-studio = "novel_studio.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/novel_studio"]
